Replace Regex with str::find
This commit builds on suggestion from @strohel
https://github.com/meilisearch/charabia/pull/181#discussion_r1117990017
and replaces the Regex matching with a `find` using a closure for these
benefits:
- We get rid of the regex dependency which, to be honest, is quite a big
  hammer for this particular problem.
- Performance boost of about 4 %. This brings the overall regression of
  implementing the camelCase splitting down to around 1 %.
- We no longer need the `unwrap` and index juggling to compute the
  boundary position.

We need a way to classify characters' unicode categories. Choose the
finl_unicode crate for it provides a nice API by extending Char and
boasts to be fastest of all currently available crates.