Fix early return condition
I initially used `== 1` because I thought that in case of splitting an
empty string we'd return an empty Iterator while in fact we'd like an
Iterator with a single value: the empty string.

I just realized that this is upheld even if we use `< 2` which just
makes much more sense here.